CN project/requirements.txt
PyQt6==6.7.1
opencv-python==4.10.0.84
numpy==2.1.2
sounddevice==0.5.0
soundfile==0.12.1
pillow==10.4.0
psutil==6.0.0
